RIVOS: Added the ability to modify DTB's and serialize them
This commit provides a way to modify a DTB using a filter_map function.
Currently only the property buffers of properties in the structure block
can be modified. Resizing of property buffers is also supported, and the
modify function will automatically move around blocks of the new dtb
to fit these new sizes.

Signed-off-by: Zayd Qumsieh <zaydq@rivosinc.com>